feat: extract Content Ops layout and sidebar
