+++ src/app/features/admin/layout/admin-sidebar.component.less
@@ -1,31 +1,3 @@
-@import '../../shared/components/header/header.component.less';
-
-.admin-layout {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.admin-header-shell {
-  padding: 0;
-  height: auto;
-  line-height: inherit;
-  background: transparent;
-  z-index: 50; // Ensure header sits above everything
-
-  // Expand standard app-header to full width in admin mode
-  .app-header__container {
-    max-width: 100% !important;
-    padding-inline: 24px;
-  }
-}
-
-.admin-body {
-  flex: 1;
-  display: flex;
-  background-color: var(--color-bg); // Use global theme bg
-}
-
 .admin-sider {
   background: var(--admin-surface-white);
   border-inline-start: 1px solid var(--color-neutral-200);
@@ -97,44 +69,9 @@
   justify-content: center;
 }
 
-.admin-content-area {
-  padding: 20px;
-  background: transparent;
-}
-
-.admin-content-inner {
-  background: var(--admin-surface-white); // can be removed later to give more space
-  padding: 16px 20px; // can be removed later to give more space
-  min-height: 380px;
-  border-radius: var(--radius-lg); // 8px // can be removed later to give more space
-  box-shadow: var(--shadow-card); // can be removed later to give more space
-  border: 1px solid var(--color-neutral-100); // can be removed later to give more space
-}
-
-.admin-header-actions {
-  display: flex;
-  gap: 16px;
-  align-items: center;
-}
-
-.admin-header-tenant {
-  flex: 1;
-  min-width: 220px;
-  max-width: 420px;
-  margin-inline: 20px;
-}
-
-.admin-header-tenant__select {
-  width: 100%;
-}
-
-.admin-header-tenant--mobile {
-  display: none;
-}
-
 // Specific override for Zorro layout alignment (RTL)
-[dir='rtl'] .ant-menu-inline.ant-menu-root .ant-menu-item,
-[dir='ltr'] .ant-menu-inline.ant-menu-root .ant-menu-item {
+:host-context([dir='rtl']) .ant-menu-inline.ant-menu-root .ant-menu-item,
+:host-context([dir='ltr']) .ant-menu-inline.ant-menu-root .ant-menu-item {
   padding-inline: 12px 16px !important;
   display: flex;
   align-items: stretch;
@@ -143,31 +80,6 @@
   & > a.admin-sidenav-link,
   & > .admin-sidenav-link {
     min-height: 38px;
-  }
-}
-
-// Adjust brand gap and toggle button for admin panel
-.admin-header-shell .app-header__brand {
-  column-gap: 16px;
-
-  .admin-desktop-toggle {
-    font-size: 24px;
-    color: var(--color-neutral-600);
-    transition: color 0.3s;
-    &:hover {
-      color: var(--color-primary-600);
-    }
-  }
-
-  .admin-mobile-toggle {
-    display: none;
-    font-size: 24px;
-    color: var(--color-neutral-600);
-    transition: color 0.3s;
-
-    &:hover {
-      color: var(--color-primary-600);
-    }
   }
 }
 
@@ -180,34 +92,6 @@
 }
 
 @media (max-width: 991px) {
-  .admin-header-shell .app-header {
-    height: auto;
-    min-height: 64px;
-  }
-
-  .admin-header-shell .app-header__container {
-    padding-block: 8px 10px;
-  }
-
-  .admin-header-tenant {
-    min-width: 0;
-    max-width: none;
-    margin-inline: 0;
-  }
-
-  .admin-header-tenant--desktop {
-    display: none;
-  }
-
-  .admin-header-tenant--mobile {
-    display: block;
-    margin-top: 8px;
-  }
-
-  .admin-header-shell .app-header__row {
-    min-height: 48px;
-  }
-
   .admin-mobile-sider-head {
     display: flex;
     justify-content: space-between;
@@ -249,16 +133,6 @@
     height: calc(100% - 50px);
   }
 
-  .admin-header-shell .app-header__brand {
-    .admin-desktop-toggle {
-      display: none;
-    }
-
-    .admin-mobile-toggle {
-      display: inline-flex;
-    }
-  }
-
   .admin-sider {
     position: fixed;
     inset-block: 0;
@@ -268,20 +142,16 @@
     box-shadow: var(--shadow-card);
   }
 
-  .admin-layout[dir='ltr'] .admin-sider {
+  :host-context([dir='ltr']) .admin-sider {
     transform: translateX(-100%);
   }
 
-  .admin-layout[dir='rtl'] .admin-sider {
+  :host-context([dir='rtl']) .admin-sider {
     transform: translateX(100%);
   }
 
   .admin-sider.admin-sider--mobile-open {
     transform: translateX(0) !important;
-  }
-
-  .admin-content-area {
-    padding: 12px;
   }
 
   .admin-mobile-backdrop {
